Handle unknown states
More gracefully handle unknown AC and Zone Status for the AirTouch 5.
This improves the behaviour of the library as the AirTouch 5 continues
to evolve and add features to reserved values in the protocol
specification.

Decoding errors for individual properties will be logged, but will not
result in the socket being disconnected. Instead it is now up to
consumers to handle these potentially unknown states.

This is a breaking change because the public API may now return None
values for several properties.